+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds `dark mode` color mode support for the `Footer`, `Header`, `HorizontalRule` and `Table` components.
 - Adds `dark mode` color mode support for the `Notification`, `ProgressIndicator`, and `SkeletonLoader` components.
 - Adds `dark mode` color mode support for the `Breadcrumbs`, `Link Types`, and `Pagination` components.
+- Adds `dark mode` color mode support for the `Accordion`, `Modal`, `Tabs`, and `Tooltip` components.
 - Adds `brand` as a `breadcrumbsType` to the `Breadcrumbs` component.
 
 ### Updates