+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Updates
 
-- Improves `TagSet` `"filter"` type accessibility by adding an `aria-label` that includes `"click to remove filter"`.
+- Improves `TagSet` `"filter"` type accessibility by adding an `aria-label` that includes `"click to remove filter"`, when `isDismissible` is true.
 
 ## 1.2.0 (October 17, 2022)