+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
   and removing the `role` attribute on the `<nav>` element.
 - Updates `AlphabetFilter` border color for the `active letter` indicator from
   `ui.gray.medium` to `ui.gray.dark`.
+- Updates background styles for `Hero` component's `primary` and `campaign` variants.
 
 ## 1.2.0 (October 17, 2022)